% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-mfa-methods.R
\name{bootStrap}
\alias{bootStrap}
\title{bootstrap}
\usage{
bootStrap(x, K = 10, L = 1000)

bootStrap(x, K = 10, L = 1000)
}
\arguments{
\item{x}{an object of class \code{"mfa"}}

\item{K}{bootstrap sample size}

\item{L}{number of bootstrap samples}
}
\description{
Bootstrap to estimate the stability of the compromise factor scores
}
\examples{
 \dontrun{
 # create a \\code{"mfa"} and plot its common factor scores
 a <- MFA()

 bootstrap(a)
 }
}

