% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-mfa-methods.R
\name{ctr.obs}
\alias{ctr.obs}
\title{contribution method}
\usage{
ctr.obs(x, obs = NULL, comp = NULL, ...)
}
\arguments{
\item{x}{an R object}

\item{obs}{which observation to analyze}

\item{comp}{which component to analyze}
}
\description{
Summarizes contribution of an observation to a dimension
}

