% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-mfa-methods.R
\name{ctr.var}
\alias{ctr.var}
\title{contribution method}
\usage{
ctr.var(x)
}
\arguments{
\item{x}{an R object}

\item{var}{which variable to analyze}

\item{comp}{which component to analyze}
}
\description{
Summarizes contribution of a variable to a dimension
}

