% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-mfa-methods.R
\name{ev.summary}
\alias{ev.summary}
\title{summary method}
\usage{
ev.summary(x)
}
\arguments{
\item{x}{an R object}

\item{digit}{number of decimal digit in print output}
}
\description{
Summarizes information about the obtained eigenvalues
}

