% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-mfa-methods.R
\name{Lg}
\alias{Lg}
\title{Lg coefficient}
\usage{
Lg(table1, table2)
}
\arguments{
\item{table1}{the first table to analyze}

\item{table2}{the second table to analyze}
}
\description{
Lg Coefficient between two tables
}

