% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-mfa-methods.R
\name{Lg_table}
\alias{Lg_table}
\title{Lg table}
\usage{
Lg_table(dataset, sets = list(1:3, 4:5, 6:10))
}
\arguments{
\item{dataset}{the dataset to study}

\item{sets}{list with sets of variables}
}
\description{
table of Lg coefficients
}

