% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{MFA}
\alias{MFA}
\alias{is.mfa}
\title{Multiple Factor Analysis}
\usage{
MFA(data = get(load("data/wine.rda")), sets = list(2:7, 8:13, 14:19, 20:24,
  25:30, 31:35, 36:39, 40:45, 46:50, 51:54), ncomps = NULL, center = TRUE,
  scale = TRUE)

is.mfa(x)
}
\arguments{
\item{data}{data set (matrix or data frame)}

\item{sets}{list of vector indicating the sets of variables (i.e. the blocks)}

\item{ncomps}{integer indicating how many number of components (i.e. factors) are to be extracted}

\item{center}{either a logical value or a numeric vector of length equal to the number of active variables in the analysis}

\item{scale}{either a logical value or a numeric vector of length equal to the number of active variables in the analysis}

\item{x}{an R object}
}
\value{
an object of class \code{"mfa"} with the following elements:

\item{eigen}{vector of eigenvalues}

\item{cfs}{matrix of common factor scores}

\item{fl}{matrix of loadings (a.k.a. factor loadings)}

\item{pfs}{matrix of partial factors scores}
}
\description{
Creates an object of class \code{"mfa"}
}
\examples{
# default
mfa1 <- MFA()

# another mfa
mfa2 <- MFA(ncomps=2)

}

