% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-mfa.R
\name{plot.mfa}
\alias{plot.mfa}
\title{Plot of object mfa}
\usage{
\method{plot}{mfa}(x, cfs = TRUE, pfs = FALSE, pfl = FALSE, num = 1,
  eig = FALSE, ...)
}
\arguments{
\item{x}{an object of class \code{"mfa"}}

\item{cfs}{whether to plot common factor scores}

\item{pfs}{whether to plot partial factor scores}

\item{pfl}{whether to plot factor loadings}

\item{num}{number of data table/block to plot}

\item{\dots}{arguments to be passed to/from other methods}
}
\description{
Plots the common factor scores and partial factor scores
}
\examples{
 \dontrun{
 # create a \\code{"mfa"} and plot its common factor scores
 a <- MFA()

 plot(a)
 }

}

