% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-mfa-methods.R
\name{RV}
\alias{RV}
\title{RV table}
\usage{
RV(table1, table2)
}
\arguments{
\item{table1}{the first table to analyze}

\item{table2}{the second table to analyze}
}
\description{
private function to evaluate the similarity between two tables
}

