% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-mfa-methods.R
\name{RV_table}
\alias{RV_table}
\title{RV table}
\usage{
RV_table(dataset, sets = list(1:3, 4:5, 6:10))
}
\arguments{
\item{sets}{list with sets of variables}
}
\description{
Coefficients to study the Between-Table Structure
}

