% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-mfa-methods.R
\name{ctr.table}
\alias{ctr.table}
\title{contribution method}
\usage{
ctr.table(x, table = NULL, comp = NULL, ...)
}
\arguments{
\item{x}{an R object}

\item{comp}{which component to analyze}

\item{var}{which table/block to analyze}
}
\description{
Summarizes contribution of a table/block to a dimension
}

